Renamed implicits instances
