+++ src/test/scala/github4s/integration/GHAuthSpec.scala
@@ -3,8 +3,8 @@ package github4s.integration
 import cats.Id
 import cats.scalatest.{ XorMatchers, XorValues }
 import github4s.Github._
-import github4s.Implicits._
-import github4s.Implicits.IdInterpreters._
+import github4s.implicits._
+import github4s.implicits.IdInterpreters._
 import github4s.Github
 import github4s.utils.TestUtils
 import org.scalatest._
