+++ src/test/scala/github4s/integration/GHReposSpec.scala
@@ -4,8 +4,8 @@ import cats.Id
 import cats.scalatest.{ XorMatchers, XorValues }
 import github4s.Github._
 import github4s.GithubResponses._
-import github4s.Implicits._
-import github4s.Implicits.IdInterpreters._
+import github4s.implicits._
+import github4s.implicits.IdInterpreters._
 import github4s.Github
 import github4s.utils.TestUtils
 import org.scalatest.{ Matchers, FlatSpec }
